Policy operations are moved to user_policy_crud.py
